New consent api and more
- setConsentData is now deprecated.
setConsentDataV2 is the new and recommended way to set manual user consent.
- Add build.yaml to configure build_version, which further on will be returned using getVersionNumber api.
- version bumps
- typos fix
