+++ .gitignore
@@ -116,4 +116,3 @@ node_modules/
 covBadgeGen.js
 coverage/
 .env
-
