+++ android/build.gradle
@@ -43,6 +43,6 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.0.0'
-    implementation 'com.appsflyer:af-android-sdk:6.15.2'
+    implementation 'com.appsflyer:af-android-sdk:6.16.2'
     implementation 'com.android.installreferrer:installreferrer:2.1'
 }
