+++ build.yaml
@@ -1,0 +1,17 @@
+targets:
+  $default:
+    builders:
+      build_version:
+        generate_for:
+          - pubspec.yaml
+        options:
+          output: lib/src/version.g.dart
+          part_of: appsflyer_sdk
+          create_field: true
+          field_name: packageVersion
+          # Use a library directive if you want to include this generated file
+          # as a part of your main library (e.g., using part 'src/version.g.dart';)
+          # By default, it generates a standalone file.
+          # part_of: appsflyer_sdk # Make it part of the main library, doesn't work properly.
+          # create_field: true # Generate the const String pluginVersion field
+          # field_name: packageVersion # Name the field
