+++ ios/Classes/AppsflyerSdkPlugin.h
@@ -18,7 +18,7 @@
 @end
 
 // Appsflyer JS objects
-#define kAppsFlyerPluginVersion             @"6.15.1"
+#define kAppsFlyerPluginVersion             @"6.16.2"
 #define afDevKey                            @"afDevKey"
 #define afAppId                             @"afAppId"
 #define afIsDebug                           @"isDebug"
