+++ lib/appsflyer_sdk.dart
@@ -6,16 +6,17 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-
+import 'package:meta/meta.dart';
 
 import 'src/callbacks.dart';
+import 'src/version.g.dart';
 
+part 'src/appsflyer_ad_revenue_data.dart';
 part 'src/appsflyer_constants.dart';
+part 'src/appsflyer_consent.dart';
 part 'src/appsflyer_invite_link_params.dart';
 part 'src/appsflyer_options.dart';
+part 'src/appsflyer_request_listener.dart';
 part 'src/appsflyer_sdk.dart';
 part 'src/udl/deep_link_result.dart';
 part 'src/udl/deeplink.dart';
-part 'src/appsflyer_consent.dart';
-part 'src/appsflyer_request_listener.dart';
-part 'src/appsflyer_ad_revenue_data.dart';
