+++ lib/src/appsflyer_sdk.dart
@@ -309,9 +309,29 @@ class AppsflyerSdk {
         "enableTCFDataCollection", {'shouldCollect': shouldCollect});
   }
 
+  @Deprecated('Use setConsentDataV2 instead')
   void setConsentData(AppsFlyerConsent consentData) {
     _methodChannel.invokeMethod('setConsentData',
         <String, dynamic>{'consentData': consentData.toMap()});
+  }
+
+  /// Sets the user consent data.
+  ///
+  /// [isUserSubjectToGDPR] - Indicates whether the user is subject to GDPR regulations.
+  /// [consentForDataUsage] - Indicates whether the user consents to data usage by AppsFlyer.
+  /// [consentForAdsPersonalization] - Indicates whether the user consents to ad personalization.
+  /// [hasConsentForAdStorage] - Indicates whether the user consents to ad storage.
+  void setConsentDataV2(
+      {bool? isUserSubjectToGDPR,
+      bool? consentForDataUsage,
+      bool? consentForAdsPersonalization,
+      bool? hasConsentForAdStorage}) {
+    _methodChannel.invokeMethod('setConsentDataV2', <String, dynamic>{
+      'isUserSubjectToGDPR': isUserSubjectToGDPR,
+      'consentForDataUsage': consentForDataUsage,
+      'consentForAdsPersonalization': consentForAdsPersonalization,
+      'hasConsentForAdStorage': hasConsentForAdStorage,
+    });
   }
 
   /// Opt-out logging for specific user
@@ -591,5 +611,11 @@ class AppsflyerSdk {
   /// Disables transfer of user-specific data over the network.
   void setDisableNetworkData(bool disable) {
     _methodChannel.invokeMethod("setDisableNetworkData", disable);
+  }
+
+  /// Retrieves the current plugin version from the generated code.
+  String getVersionNumber() {
+    // Use the constant generated by build_version
+    return packageVersion;
   }
 }
