+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: appsflyer_sdk
 description: A Flutter plugin for AppsFlyer SDK. Supports iOS and Android.
-version: 6.15.2
+version: 6.16.2
 
 homepage: https://github.com/AppsFlyerSDK/flutter_appsflyer_sdk
 
@@ -18,6 +18,8 @@ dev_dependencies:
   test: ^1.16.5
   mockito: ^5.4.4
   effective_dart: ^1.3.0
+  build_runner: ^2.4.11
+  build_version: ^2.1.1
 
 
 flutter:
